image_generation/requirements.txt
torch==2.5.0
torchvision==0.20.0
torchaudio==2.5.0
diffusers==0.32.2
transformers==4.48.0
open3d==0.19.0
pillow==11.1.0
numpy==1.24.0
accelerate==0.28.0


